Create index.html
This is my Brick and Mortar Bitcoin Business Index. The idea is to help foster a bitcoin circular economy.  I hope Bitcoiner's will look at this list to find businesses that accept bitcoin as payment.

If you like this idea, please contribute. This website needs input from you.  Know of a coffee shop that accepts bitcoin in Alaska or a computer repair shop that accepts bitcoin in Rugby North Dakota?  Please add it to the list.  Thank you,